:recycle: :art: Separación de GraphInErrorCase en otro componenete para incrementar la reutilizacion de codigo
